chore: move CallbackToken, Social::Nonce and cereal::NullType to their own headers
